Remove unused logic

+++ build.sh
@@ -56,13 +56,13 @@ declare -r linkflags='-Xlinker -s'
 
 declare -ra targets=(
 	'aarch64-unknown-linux-android'
-	'riscv64-unknown-linux-android'
-	'mipsel-unknown-linux-android'
-	'i686-unknown-linux-android'
-	'armv7-unknown-linux-androideabi'
-	'x86_64-unknown-linux-android'
-	'armv5-unknown-linux-androideabi'
-	'mips64el-unknown-linux-android'
+	# 'riscv64-unknown-linux-android'
+	# 'mipsel-unknown-linux-android'
+	# 'i686-unknown-linux-android'
+	# 'armv7-unknown-linux-androideabi'
+	# 'x86_64-unknown-linux-android'
+	# 'armv5-unknown-linux-androideabi'
+	# 'mips64el-unknown-linux-android'
 )
 
 declare -ra versions=(
@@ -917,9 +917,12 @@ for triplet in "${targets[@]}"; do
 		cd "${sysroot_directory}/lib"
 		
 		echo 'GROUP ( ../libm.so AS_NEEDED ( ../libm.a ) )' > './ldscripts/libm.so'
-		echo 'GROUP ( ../libc.so AS_NEEDED ( ../libc.a ) )' > './ldscripts/libc.so'
 		
-		mkdir 'gcc' 'static' 'no-lfs'
+		if [ -f './libc.a' ]; then
+			echo 'GROUP ( ../libc.so AS_NEEDED ( ../libc.a ) )' > './ldscripts/libc.so'
+		fi
+		
+		mkdir 'gcc' 'static'
 		
 		ln --symbolic --relative './lib'*'.'{so,a} './static'
 		ln --symbolic --relative './crt'*'.o' './static'
@@ -954,29 +957,6 @@ for triplet in "${targets[@]}"; do
 			ln --symbolic "${library}" './'
 			ln --symbolic --relative "${library}" './gcc'
 		done
-		
-		if (( ! abi64 && version < 24 )); then
-			ln --symbolic --relative './'*  './no-lfs'
-			
-			unlink './no-lfs/no-lfs'
-			unlink './no-lfs/static'
-			mkdir './no-lfs/static'
-			
-			ln --symbolic --relative './static/'* './no-lfs/static'
-			
-			unlink './no-lfs/libc.so'
-			cp './libc.so' './no-lfs'
-			
-			sed \
-				--in-place \
-				--expression 's/fgetpos/ae874db/g' \
-				--expression 's/fsetpos/f94d1c7/g' \
-				--expression 's/fseeko/c690a7/g' \
-				--expression 's/ftello/deb0b1/g' \
-				'./no-lfs/libc.so'
-			
-			ln --symbolic --relative --force './no-lfs/libc.so' './no-lfs/static/libc.so'
-		fi
 		
 		cp "${gcc_wrapper}" "${toolchain_directory}/bin/${triplet}${version}-gcc"
 		cp "${gcc_wrapper}" "${toolchain_directory}/bin/${triplet}${version}-g++"
